requirements.txt
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
dill==0.3.9
distlib==0.3.9
docutils==0.20.1
filelock==3.16.1
identify==2.6.1
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
isort==5.13.2
jaraco.classes==3.3.1
keyring==24.3.0
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.2.0
nh3==0.2.15
nodeenv==1.9.1
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pkginfo==1.9.6
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.17.2
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
readme-renderer==42.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.0
setuptools==69.1.0
six==1.16.0
tomlkit==0.13.2
twine==5.0.0
tzdata==2024.2
urllib3==2.2.1
virtualenv==20.27.0
wheel==0.42.0
zipp==3.17.0
